Gas Optimizations (#28)
* Gas optimizations for batch CTF operations

- Batch mint/merge operations instead of per-maker calls
- Batch fee transfers for SELL orders
- Storage packing: OrderStatus.remaining as uint248
- Add gas snapshot tests (JSON format)

* snapshot

+++ src/exchange/libraries/Structs.sol
@@ -100,5 +100,5 @@ enum MatchType {
 
 struct OrderStatus {
     bool filled;
-    uint256 remaining;
+    uint248 remaining;
 }
